Create Telescope picker for Go pkg symbols
The definition of a "workspace" is not quite the same between gopls and
LSP. Therefore, the Telescope lsp_workspace_symbols picker requires a
query in order to scope down; no query returns no results, not all
symbols in the `go.work` defined modules.

Use some vim shorthands to get the package name of the current file,
then only show the symbols with that. Super helpful for seeing all
symbols within a Go package when they're spread across files.

Signed-off-by: Nolan Brubaker <nolan@nbrubaker.com>